Revert "fix: перенес статистику"
