+++ src/features/dashboard/index.ts
@@ -1,0 +1,1 @@
+export { useDashboardStatistics } from './lib/use-dashboard-statistics'
